refac: Moved some generic classes to BuildingBlocks
- Renamed Position class to Location to be more generic
